# Conflicts:
#	backend/target/backend-0.0.1-SNAPSHOT.jar.original
#	backend/target/classes/resources/main-es2015.js
#	backend/target/classes/resources/main-es2015.js.map
#	backend/target/classes/resources/main-es5.js
#	backend/target/classes/resources/main-es5.js.map
#	frontend/target/classes/static/main-es2015.js
#	frontend/target/classes/static/main-es2015.js.map
#	frontend/target/classes/static/main-es5.js
#	frontend/target/classes/static/main-es5.js.map
#	sqlitesample.db

+++ backend/src/main/java/com/projetM2_foot/api/dto/EntiteDto.java
@@ -36,11 +36,11 @@ public class EntiteDto {
             description = "Position initial X de l'entité",
             example = "5"
     )
-    private int x;
+    private float x;
 
     @Schema(
             description = "Position initial Y de l'entité",
             example = "3"
     )
-    private int y;
+    private float y;
 }
