+++ backend/src/main/java/com/projetM2_foot/api/request/DeplacementRequestCreate.java
@@ -14,9 +14,9 @@ public class DeplacementRequestCreate {
     private int numScene;
     private int numBloc;
 
-    private int startPosX;
-    private int startPosY;
-    private int endPosX;
-    private int endPosY;
+    private float startPosX;
+    private float startPosY;
+    private float endPosX;
+    private float endPosY;
 
 }
