+++ backend/src/main/java/com/projetM2_foot/api/request/EntiteRequestCreate.java
@@ -1,13 +1,8 @@
 package com.projetM2_foot.api.request;
 
-import com.projetM2_foot.entity.Scenario;
+
 import lombok.Builder;
 import lombok.Data;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.validation.constraints.NotNull;
 
 @Data
 @Builder
@@ -16,7 +11,7 @@ public class EntiteRequestCreate {
     private Long scenario;
     private int type;
     private int numero;
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
 }
